Move method, make stuff static
